requirements.txt
certifi==2019.11.28
chardet==3.0.4
future==0.18.2
get==2019.4.13
idna==2.8
oauthlib==3.1.0
pkg-resources==0.0.0
post==2019.4.13
public==2019.4.13
python-twitter==3.5
query-string==2019.4.13
request==2019.4.13
requests==2.22.0
requests-oauthlib==1.3.0
urllib3==1.25.7
